# Optimize `cut_into_windows` for long cuts