@pzelasko 2023-09-19
> @pzelasko I do not see particular benefits of keeping limited to Python built-in types. Tbh, I would rather migrate everything related to manifests and their sets into some kind of performant pure immutable persistent trees and sequences with laziness by default (i.e. rewrite everything in Haskell lol).

At some point I contemplated moving basic data types to C++ structs for smaller memory footprint and faster (de)serialization, and maybe actual immutability, but it would have drastically complicated maintenance and further development of the library and limited the number of people who could help out. I like the idea in principle but it's likely not worth the trouble.

> To keep it real, I think it would be quite nice to have a kind of `Splittable` interface with methods `split_right` and `split_left`, which are always <= O(log n). Both methods split the interval-based object (and the set of all its descendants) into two around the point and returns a pair of left and right objects. If the object is in fact atomic (i.e. a `Supervision` without alignments and with text or custom data), it is a no-op, and the original object is returned either as left or right result (depending on the method called).
> 
> Split is more simple and fundamental than `truncate`, and `truncate` and all other operations can be implemented using it. Some operations (`cut_into_windows` being the example) are implemented more efficiently using individual splits (and also in a potentially easily parallelizable way, if we cut dichotomically starting from the middle).
> 
> If any type of manifest supports `Splittable` efficiently, then doing all other operations efficiently seems almost trivial. We can define proper `__repr__`, `__getitem__` and `__iter__` for our generic IntervalTree based container which supports splits so that it can be treated completely similarly to a simple list, if a user wants. I do not see particular upsides in keeping strictly to the builtin data structures.

I understand the technical solution, but what would be the impact/motivation?