@flyingleafe 2023-09-20
@pzelasko The motivation is to unify and heavily simplify all logic around cutting stuff recursively, throwing out large amounts of repetitive code and stopping caring about performance optimizations on a case-by-case basis. I.e. imagine if `MixedCut` was like
```
class MixedCut(Cut):
     id: str
     tracks: OurIntervalTree[MixTrack]
     transforms: ...
     
     ...
```
and `Cut` is supposed to implement `Splittable`, and `OurIntervalTree` is `Splittable`, such that it continues to split a border element further if the element is splittable itself, and `MixTrack` is trivially also `Splittable`, being just a thin wrapper around `Cut`. A correct `split` method implementation (from which `truncate` is derived), would then be just:
```
def split(self, offset: Seconds, side: Side = LEFT, preserve_id: Optional[Side] = None) -> Tuple[Cut, Cut]:
    l_tracks, r_tracks = self.tracks.split(offset, side, preserve_id)
    l_id = self.id if preserve_id == LEFT else str(uuid4())
    r_id = self.id if preserve_id == RIGHT else str(uuid4())
    return (
        self._fastcopy(id=l_id, tracks=l_tracks, duration=offset),
        self._fastcopy(id=r_id, tracks=r_tracks, duration=self.duration - offset)
    )
```
where `_fastcopy` handles the edge cases of no tracks left and SNR fix:
```
def _fastcopy(self, id, tracks, duration=None):
   np_tracks = [t for t in tracks if not isinstance(t, PaddingCut)]
   duration = self.duration if duration is None else duration
   if len(np_tracks) == 0:
      return PaddingCut(id=id, sampling_rate=self.sampling_rate, duration=duration, ...)
   if all(t.snr is not None for t in np_tracks):
      tracks = ... //... do this snr resetting
   return fastcopy(self, tracks=tracks)
```

For the `DataCut` with supervisions inside an interval tree, `split` implementation would look almost exactly the same. Well, actually, it can be even exactly the same and be shared inside the `Cut` definition, if we:
- abstract away general `fastcopy` as an overridable method;
- abstract away a list of objects' properties which should be `split` further (e.g. `supervisions` for `DataCut`, `tracks` for `MixedCut`);
- handle `start` and `duration` properly for `DataCut`, ignoring them in `MixedCut`s implementation of `fastcopy`.

Imagine how much code could be thrown out and how much smaller the codebase could become.
So basically what I say is that such an approach can be a starting point in a major (yet very doable) refactoring which would result in Lhotse's codebase becoming much smaller, more maintainable and _more_ performant.